Split out nav stories into separate files
